requirements.txt

sacrebleu
rouge_score
py7zr
pandas
nltk
tqdm
PyYaml
matplotlib
-f https://download.pytorch.org/whl/cu126
torch
transformers
transformers[sentencepiece]
datasets
python-box
ensure
hf_xet
evaluate
#-e .